define :meth:`InKernelCallable.get_called_callables`
