+++ loopy/kernel/tools.py
@@ -36,6 +36,10 @@ from loopy.kernel import LoopKernel
 from loopy.translation_unit import (TranslationUnit,
                                     for_each_kernel)
 from loopy.kernel.function_interface import CallableKernel
+from loopy.kernel.instruction import (
+        MultiAssignmentBase, CInstruction, _DataObliviousInstruction)
+from loopy.symbolic import CombineMapper
+from functools import reduce
 import logging
 logger = logging.getLogger(__name__)
 
@@ -1992,6 +1996,60 @@ def infer_args_are_input_output(kernel):
         new_args.append(arg)
 
     return kernel.copy(args=new_args)
+
+# }}}
+
+
+# {{{ CallablesIDCollector
+
+class CallablesIDCollector(CombineMapper):
+    """
+    Mapper to collect function identifiers of all resolved callables in an
+    expression.
+    """
+    def combine(self, values):
+        import operator
+        return reduce(operator.or_, values, frozenset())
+
+    def map_resolved_function(self, expr):
+        return frozenset([expr.name])
+
+    def map_constant(self, expr):
+        return frozenset()
+
+    def map_kernel(self, kernel):
+        callables_in_insn = frozenset()
+
+        for insn in kernel.instructions:
+            if isinstance(insn, MultiAssignmentBase):
+                callables_in_insn = callables_in_insn | (
+                        self(insn.expression))
+            elif isinstance(insn, (CInstruction, _DataObliviousInstruction)):
+                pass
+            else:
+                raise NotImplementedError(type(insn).__name__)
+
+        for rule in kernel.substitutions.values():
+            callables_in_insn = callables_in_insn | (
+                    self(rule.expression))
+
+        return callables_in_insn
+
+    def map_type_cast(self, expr):
+        return self.rec(expr.child)
+
+    map_variable = map_constant
+    map_function_symbol = map_constant
+    map_tagged_variable = map_constant
+
+
+def get_resolved_callable_ids_called_by_knl(knl, callables):
+    clbl_id_collector = CallablesIDCollector()
+    callables_called_by_kernel = clbl_id_collector.map_kernel(knl)
+    callables_called_by_called_callables = frozenset().union(*(
+        callables[clbl_id].get_called_callables(callables)
+        for clbl_id in callables_called_by_kernel))
+    return callables_called_by_kernel | callables_called_by_called_callables
 
 # }}}
