+++ loopy/translation_unit.py
@@ -27,18 +27,15 @@ from pymbolic.primitives import Variable
 from functools import wraps
 
 from loopy.symbolic import (RuleAwareIdentityMapper, ResolvedFunction,
-        CombineMapper, SubstitutionRuleMappingContext)
+                            SubstitutionRuleMappingContext)
 from loopy.kernel.function_interface import (
         CallableKernel, ScalarCallable)
-from loopy.kernel.instruction import (
-        MultiAssignmentBase, CInstruction, _DataObliviousInstruction)
 from loopy.diagnostic import LoopyError
 from loopy.library.reduction import ReductionOpFunction
 
 from loopy.kernel import LoopKernel
 from loopy.tools import update_persistent_hash
 from pymbolic.primitives import Call
-from functools import reduce
 from pyrsistent import pmap, PMap
 
 __doc__ = """
@@ -411,70 +408,13 @@ def rename_resolved_functions_in_a_single_kernel(kernel,
 # }}}
 
 
-# {{{ CallablesIDCollector
-
-class CallablesIDCollector(CombineMapper):
+def get_reachable_resolved_callable_ids(callables, entrypoints):
     """
-    Mapper to collect function identifiers of all resolved callables in an
-    expression.
+    Returns a :class:`frozenset` of callables ids that are resolved and
+    reachable from *entrypoints*.
     """
-    def combine(self, values):
-        import operator
-        return reduce(operator.or_, values, frozenset())
-
-    def map_resolved_function(self, expr):
-        return frozenset([expr.name])
-
-    def map_constant(self, expr):
-        return frozenset()
-
-    def map_kernel(self, kernel):
-        callables_in_insn = frozenset()
-
-        for insn in kernel.instructions:
-            if isinstance(insn, MultiAssignmentBase):
-                callables_in_insn = callables_in_insn | (
-                        self(insn.expression))
-            elif isinstance(insn, (CInstruction, _DataObliviousInstruction)):
-                pass
-            else:
-                raise NotImplementedError(type(insn).__name__)
-
-        for rule in kernel.substitutions.values():
-            callables_in_insn = callables_in_insn | (
-                    self(rule.expression))
-
-        return callables_in_insn
-
-    def map_type_cast(self, expr):
-        return self.rec(expr.child)
-
-    map_variable = map_constant
-    map_function_symbol = map_constant
-    map_tagged_variable = map_constant
-
-
-def _get_reachable_callable_ids_for_knl(knl, callables):
-    clbl_id_collector = CallablesIDCollector()
-
-    def rec(clbl_id):
-        clbl = callables[clbl_id]
-        if isinstance(clbl, CallableKernel):
-            return (_get_reachable_callable_ids_for_knl(clbl.subkernel, callables)
-                    | frozenset([clbl_id]))
-        else:
-            return frozenset([clbl_id])
-
-    return frozenset().union(*(rec(clbl_id)
-                               for clbl_id in clbl_id_collector.map_kernel(knl)))
-
-
-def _get_reachable_callable_ids(callables, entrypoints):
-    return frozenset().union(*(
-        _get_reachable_callable_ids_for_knl(callables[e].subkernel, callables)
-        for e in entrypoints))
-
-# }}}
+    return frozenset().union(*(callables[e].get_called_callables(callables)
+                               for e in entrypoints))
 
 
 # {{{ CallablesInferenceContext
@@ -631,8 +571,8 @@ class CallablesInferenceContext(ImmutableRecord):
         # {{{ get all the callables reachable from the new entrypoints.
 
         # get the names of all callables reachable from the new entrypoints
-        new_callable_ids = _get_reachable_callable_ids(
-                self.callables, self.new_entrypoints)
+        new_callable_ids = get_reachable_resolved_callable_ids(self.callables,
+                                                               self.new_entrypoints)
 
         # get the history of function ids from the performed renames:
         history = {}
